requirements.txt
python-dotenv
requests
fastapi
uvicorn[standard]
python-multipart 